Added links to a beginner-friendly repo + beginner's checklist
The CockroachDB To-Do apps repo is a beginner-friendly repo to enable participants to build CockroachDB To-Do apps using a language of their preference. The intention here is to let participants choose a language they are comfortable with as well as learn a new technology/skill (building CRUD apps) and helps CockroachDB build our sample apps repo. 

Also added a link to a checklist for absolute beginners.